Update StudentController.php
When students register for CTT, the first step is to verify their eligibility. This verification process involves comparing their exam scores with the minimum marks specified in the eligibility table. After this verification, the system informs the user whether they meet the eligibility criteria or not. If the students are found to be eligible, their contact numbers are recorded, and their eligibility status is updated to "true." In cases where they do not meet the eligibility criteria, their eligibility status is updated to "false."